feat: 쿠폰 디테일 빌드 오류 수정
